added preview
