+++ app/src/main/java/com/example/todobarzh/ui/screens/common/ShadowTopAppBarModifier.kt
@@ -1,4 +1,4 @@
-package com.example.todobarzh.ui.components.common
+package com.example.todobarzh.ui.screens.common
 
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.TopAppBarState
